dashboard: hide heartbeats from RPC not in guardian set
